PmdMon::load took a uint8_t, so Hoenn drew Kanto sprites
Reported from a board: a MARSHTOMP appeared as an IVYSAUR. 258 - 256 = 2.

PmdMon::load() and SdMon::load() both took a uint8_t dexNum while the
dex had grown to 386, so every species from 256 up wrapped into Kanto and
opened the wrong file -- 131 species, the whole back half of Hoenn. The
battle path even had an explicit (uint8_t) cast on the way in.

This is the THIRD time this trap has been sprung by the same expansion:
DexEntry::evolvesTo and TrainerMon::dex were both uint8_t too. The
thumbnails were never affected, because SdThumbs::get() takes an int16_t
-- which is why the gallery looked right while the creature on the main
screen was somebody else, and why this survived so long.

Both loaders take int16_t now and refuse a dex outside 1..999 rather than
opening some other creature's file. PmdMon records which species it holds.

sprite_test loads all 386 off the sprite directory. The assertion that
matters compares the PIXELS of every species past 255 against the Kanto
species it would truncate onto -- because m.dex only records what was
ASKED for, and a truncation hidden inside the filename slipped past that
check when tried. Both sabotages fail it now.

+++ TamaPoke.ino
@@ -2876,7 +2876,7 @@ static void btlSyncSprite(uint8_t who, const Combatant &c) {
   btlPmd[who].unload();
   btlPmdDex[who] = 0;
   if (c.dex < 1 || c.dex > DEX_COUNT) return;
-  if (btlPmd[who].load((uint8_t)c.dex, c.shiny)) btlPmdDex[who] = key;
+  if (btlPmd[who].load(c.dex, c.shiny)) btlPmdDex[who] = key;   // NOT (uint8_t): Hoenn runs past 255
 }
 
 static void btlFreeSprites() {
